zigbee: fixed java mapping of Bitmap and General data types.

+++ org.osgi.service.zigbee/src/org/osgi/service/zigbee/types/ZigBeeBitmap16.java
@@ -52,7 +52,7 @@ public class ZigBeeBitmap16
 	}
 
 	public Class getJavaDataType() {
-		return Integer.class;
+		return Short.class;
 	}
 
 	public short getId() {
