+++ org.osgi.service.zigbee/src/org/osgi/service/zigbee/types/ZigBeeGeneralData16.java
@@ -52,7 +52,7 @@ public class ZigBeeGeneralData16
 	}
 
 	public Class getJavaDataType() {
-		return Integer.class;
+		return Short.class;
 	}
 
 	public short getId() {
